+++ packages/api/src/process-lcd.ts
@@ -1,5 +1,6 @@
 import * as ProcessType from './typedef/process'
 import LCDClient from './util/lcd'
+import { IMsg } from './transaction'
 
 export type IResult = {
   instanceHash: string;
@@ -16,8 +17,102 @@ export type ITask = {
   taskKey: string;
 }
 
-export type IMap = any
-export type IFilter = any
+export type IOutputNullType = {
+  type: 'mesg.types.Process_Node_Map_Output_Null_';
+  value: {
+    null?: 0
+  }
+}
+
+export type IOutputStringType = {
+  type: 'mesg.types.Process_Node_Map_Output_StringConst';
+  value: {
+    string_const: string;
+  }
+}
+
+export type IOutputDoubleType = {
+  type: 'mesg.types.Process_Node_Map_Output_DoubleConst';
+  value: {
+    double_const: number;
+  }
+}
+
+export type IOutputBoolType = {
+  type: 'mesg.types.Process_Node_Map_Output_BoolConst';
+  value: {
+    bool_const: boolean;
+  }
+}
+
+export type IOutputListType = {
+  type: 'mesg.types.Process_Node_Map_Output_List_';
+  value: {
+    list: {
+      outputs: IOutput[]
+    }
+  }
+}
+
+export type IOutputMapType = {
+  type: 'mesg.types.Process_Node_Map_Output_Map_';
+  value: {
+    map: IMapOutput[]
+  }
+}
+
+export type IRefSelectorKey = {
+  type: "mesg.types.Process_Node_Map_Output_Reference_Path_Key";
+  value: {
+    key: string;
+  }
+}
+
+export type IRefSelectorIndex = {
+  type: "mesg.types.Process_Node_Map_Output_Reference_Path_Index";
+  value: {
+    index?: string;
+  }
+}
+
+export type IRefPath = {
+  path?: IRefPath
+  Selector: IRefSelectorKey | IRefSelectorIndex
+}
+
+export type IOutputRefType = {
+  type: 'mesg.types.Process_Node_Map_Output_Ref';
+  value: {
+    ref: {
+      nodeKey: string;
+      path: IRefPath
+    }
+  }
+}
+
+export type IOutput = {
+  Value: IOutputNullType | IOutputStringType | IOutputDoubleType | IOutputBoolType | IOutputListType | IOutputMapType | IOutputRefType
+}
+
+export type IMapOutput = {
+  Key: string,
+  Value: IOutput
+}
+
+export enum FilterPredicate {
+  Unknown = 0,
+  EQ = 1
+}
+
+export type IFilterCondition = {
+  key: string;
+  predicate: FilterPredicate;
+  value: string;
+}
+
+export type IFilter = {
+  conditions: IFilterCondition[];
+}
 
 export type IResultType = {
   type: 'mesg.types.Process_Node_Result_';
@@ -43,7 +138,7 @@ export type ITaskType = {
 export type IMapType = {
   type: 'mesg.types.Process_Node_Map_';
   value: {
-    map: IMap;
+    map: IMapOutput[];
   }
 }
 
@@ -72,7 +167,42 @@ export type IProcessRequest = {
   edges: ProcessType.mesg.types.Process.IEdge[];
 }
 
+export type IMsgCreate = {
+  address: string;
+  request: IProcessRequest;
+}
+
+export type IMsgDelete = {
+  address: string;
+  request: {
+    hash: string;
+  }
+}
+
 export default class Process extends LCDClient {
+
+  createMsg(owner: string, process: IProcessRequest): IMsg<IMsgCreate> {
+    return {
+      type: 'process/CreateProcess',
+      value: {
+        address: owner,
+        request: process
+      }
+    }
+  }
+
+  deleteMsg(owner: string, hash: string): IMsg<IMsgDelete> {
+    return {
+      type: 'process/DeleteProcess',
+      value: {
+        address: owner,
+        request: {
+          hash
+        }
+      }
+    }
+  }
+
   async get(hash: string): Promise<IProcess> {
     return (await this.query(`/process/get/${hash}`)).result
   }