+++ packages/compiler/CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Breaking Changes
 
 - [#186](https://github.com/mesg-foundation/js-sdk/pull/186) Compile process in a LCD format
+- [#188](https://github.com/mesg-foundation/js-sdk/pull/188) Compile services in a LCD format
 
 #### Improvements
 #### Bug fixes